✨ feat: adicionar componentes de cabeçalho e manipulação de modal, além de validação de votação
